+++ demo-shell-ng2/systemjs.config.js
@@ -18,7 +18,7 @@
         'ng2-alfresco-search': 'node_modules/ng2-alfresco-search/dist',
         'ng2-alfresco-upload': 'node_modules/ng2-alfresco-upload/dist',
         'ng2-alfresco-viewer': 'node_modules/ng2-alfresco-viewer/dist',
-        'ng2-alfresco-activiti-form': 'node_modules/ng2-alfresco-activiti-form/dist'
+        'ng2-activiti-form': 'node_modules/ng2-activiti-form/dist'
     };
     // packages tells the System loader how to load when no filename and/or no extension
     var packages = {
@@ -35,7 +35,7 @@
         'ng2-alfresco-search': { main: 'index.js', defaultExtension: 'js'},
         'ng2-alfresco-upload': { main: 'index.js', defaultExtension: 'js'},
         'ng2-alfresco-viewer': { main: 'index.js', defaultExtension: 'js'},
-        'ng2-alfresco-activiti-form': { main: 'index.js', defaultExtension: 'js'}
+        'ng2-activiti-form': { main: 'index.js', defaultExtension: 'js'}
     };
     var ngPackageNames = [
         'common',
