+++ ng2-components/ng2-activiti-form/.editorconfig
