+++ ng2-components/ng2-activiti-form/assets/license_header.txt
