+++ ng2-components/ng2-activiti-form/karma-test-shim.js
