+++ ng2-components/ng2-activiti-form/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "ng2-alfresco-activiti-form",
-  "description": "Alfresco Angular2 Activiti Form Component",
+  "name": "ng2-activiti-form",
+  "description": "Alfresco Activiti Form Component for Angular 2",
   "version": "0.2.0",
   "author": "Alfresco Software, Ltd.",
   "scripts": {
