+++ ng2-components/ng2-activiti-form/src/components/widgets/index.ts
