+++ ng2-components/ng2-activiti-form/src/components/widgets/tabs/tabs.widget.ts
